+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 5. Shape [stroke](https://svgwg.org/svg2-draft/painting.html#StrokeProperties) support
 6. [`viewBox`](https://svgwg.org/svg2-draft/coords.html#ViewBoxAttribute) support on the [`<svg>`](https://svgwg.org/svg2-draft/struct.html#SVGElement) element
 7. Support the [`circle`](https://svgwg.org/svg2-draft/shapes.html#CircleElement) element
+8. Support the [`ellipse`](https://svgwg.org/svg2-draft/shapes.html#EllipseElement) element
 
 ### Performance improvements
 1. `ColorMaterialRenderer` and `TextureMaterialRenderer` are now globally shared instead of per `Artwork`, which previously made the creation time of tiles significantly slower.